refactor: complex도메인 분리
