+++ src/main/java/com/kb/zipkim/domain/prop/service/PropertyService.java
@@ -3,12 +3,12 @@ package com.kb.zipkim.domain.prop.service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.kb.zipkim.domain.prop.dto.*;
-import com.kb.zipkim.domain.prop.entity.Complex;
+import com.kb.zipkim.domain.complex.entity.Complex;
 import com.kb.zipkim.domain.prop.entity.Property;
 import com.kb.zipkim.domain.prop.file.FileStoreService;
 import com.kb.zipkim.domain.prop.file.UploadFile;
 import com.kb.zipkim.domain.prop.repository.ComplexPropQueryRepository;
-import com.kb.zipkim.domain.prop.repository.ComplexRepository;
+import com.kb.zipkim.domain.complex.repository.ComplexRepository;
 import com.kb.zipkim.domain.prop.repository.PropertyRepository;
 import com.kb.zipkim.domain.register.entity.Registered;
 import com.kb.zipkim.domain.register.repository.RegisteredRepository;
@@ -35,20 +35,15 @@ import java.util.List;
 @RequiredArgsConstructor
 public class PropertyService {
 
-    private static final String KEY = "complexes";
     private final PropertyRepository propertyRepository;
     private final ComplexRepository complexRepository;
     private final FileStoreService fileStoreService;
-    private final ObjectMapper objectMapper;
-
-    private final RedisTemplate<String,String> redisTemplate;
     private final ComplexPropQueryRepository complexPropQueryRepository;
     private final RegisteredRepository registeredRepository;
 
     @Transactional
     public RegisterResult registerProp(PropRegisterForm form) throws IOException {
         List<UploadFile> uploadFiles = fileStoreService.storeFiles(form.getImages());
-        GeoOperations<String, String> geoOperations = redisTemplate.opsForGeo();
 
         Property property = Property.makeProperty(form);
 
@@ -76,33 +71,6 @@ public class PropertyService {
         }
         result.setPropId(property.getId());
         return result;
-    }
-
-    public List<NearByComplex> findNearComplexes(String type, double latitude, double longitude, double radius) throws JsonProcessingException {
-        GeoOperations<String, String> geoOperations = redisTemplate.opsForGeo();
-        GeoReference<String> reference = GeoReference.fromCoordinate(new Point(longitude, latitude));
-        Distance distance = new Distance(radius, RedisGeoCommands.DistanceUnit.KILOMETERS);
-
-        RedisGeoCommands.GeoRadiusCommandArgs args = RedisGeoCommands.GeoRadiusCommandArgs.newGeoRadiusArgs()
-                .includeCoordinates()
-                .includeDistance()
-                .sortAscending();
-
-        GeoResults<RedisGeoCommands.GeoLocation<String>> results = geoOperations.search(KEY, reference, distance, args);
-        List<NearByComplex> complexes = new ArrayList<>();
-
-        if(results == null) return complexes; //주변값이 없으면 빈 배열반환
-
-        for (GeoResult<RedisGeoCommands.GeoLocation<String>> result : results) {
-            RedisGeoCommands.GeoLocation<String> location = result.getContent();
-            NearByComplex complex = objectMapper.readValue(location.getName(), NearByComplex.class);
-            if (!complex.getType().equals(type) ) {
-                continue;
-            }
-            complex.setDistance(result.getDistance().getValue());
-            complexes.add(complex);
-        }
-        return complexes;
     }
 
     public Page<SimplePropInfo> findPropList(Long complexId, Pageable pageable) {
