+++ src/main/java/com/kb/zipkim/global/redis/CacheWarmingService.java
@@ -1,15 +1,12 @@
-package com.kb.zipkim.domain.prop.redis;
+package com.kb.zipkim.global.redis;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.kb.zipkim.domain.prop.dto.NearByComplex;
-import com.kb.zipkim.domain.prop.entity.Complex;
-import com.kb.zipkim.domain.prop.entity.Property;
-import com.kb.zipkim.domain.prop.repository.ComplexRepository;
-import com.kb.zipkim.domain.prop.repository.PropertyRepository;
+import com.kb.zipkim.domain.complex.dto.NearByComplex;
+import com.kb.zipkim.domain.complex.entity.Complex;
+import com.kb.zipkim.domain.complex.repository.ComplexRepository;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.geo.Point;
 import org.springframework.data.redis.core.GeoOperations;
